Prevent server sound ear destruction / Particle effect rand param
