+++ core/src/mindustry/core/NetClient.java
@@ -164,14 +164,14 @@ public class NetClient implements ApplicationListener{
     public static void sound(Sound sound, float volume, float pitch, float pan){
         if(sound == null) return;
 
-        sound.play(volume * Core.settings.getInt("sfxvol") / 100f, pitch, pan);
+        sound.play(Mathf.clamp(volume, 0, 2f) * Core.settings.getInt("sfxvol") / 100f, pitch, pan);
     }
 
     @Remote(variants = Variant.both, unreliable = true)
     public static void soundAt(Sound sound, float x, float y, float volume, float pitch){
         if(sound == null) return;
 
-        sound.at(x, y, pitch, volume);
+        sound.at(x, y, pitch, Mathf.clamp(volume, 0, 2f));
     }
 
     @Remote(variants = Variant.both, unreliable = true)
