+++ core/src/mindustry/world/blocks/production/Fracker.java
@@ -1,18 +1,11 @@
 package mindustry.world.blocks.production;
 
-import arc.graphics.g2d.*;
 import arc.math.*;
-import mindustry.annotations.Annotations.*;
 import mindustry.gen.*;
-import mindustry.graphics.*;
 import mindustry.world.meta.*;
 
 public class Fracker extends SolidPump{
     public float itemUseTime = 100f;
-
-    public @Load("@-liquid") TextureRegion liquidRegion;
-    public @Load("@-rotator") TextureRegion rotatorRegion;
-    public @Load("@-top") TextureRegion topRegion;
 
     public Fracker(String name){
         super(name);
@@ -30,32 +23,8 @@ public class Fracker extends SolidPump{
         stats.add(Stat.productionTime, itemUseTime / 60f, StatUnit.seconds);
     }
 
-    @Override
-    public boolean outputsItems(){
-        return false;
-    }
-
-    @Override
-    public TextureRegion[] icons(){
-        return new TextureRegion[]{region, rotatorRegion, topRegion};
-    }
-
     public class FrackerBuild extends SolidPumpBuild{
         public float accumulator;
-
-        @Override
-        public void drawCracks(){}
-
-        @Override
-        public void draw(){
-            Draw.rect(region, x, y);
-            super.drawCracks();
-
-            Drawf.liquid(liquidRegion, x, y, liquids.get(result) / liquidCapacity, result.color);
-
-            Drawf.spinSprite(rotatorRegion, x, y, pumpTime);
-            Draw.rect(topRegion, x, y);
-        }
 
         @Override
         public void updateTile(){
@@ -72,11 +41,6 @@ public class Fracker extends SolidPump{
                 lastPump = 0f;
                 dumpLiquid(result);
             }
-        }
-
-        @Override
-        public float typeLiquid(){
-            return liquids.get(result);
         }
     }
 }
